Separate out coordinate map tests
